+++ jest.config.cjs
@@ -1,0 +1,12 @@
+module.exports = {
+  testEnvironment: 'node',
+  transform: {
+    '^.+\\.jsx?$': 'babel-jest',
+  },
+  moduleFileExtensions: ['js', 'json'],
+  testMatch: ['**/tests/**/*.test.js'],
+  coverageDirectory: './coverage',
+  collectCoverage: true,
+  collectCoverageFrom: ['src/**/*.js'],
+  coverageReporters: ['text', 'lcov'],
+};