+++ src/application.js
@@ -5,6 +5,8 @@ import routes from './routes/index.js';
 
 const app = express();
 app.use(cors())
+app.use(express.json({ limit: '100mb' }));
+app.use(express.urlencoded({ limit: '100mb', extended: true }));
 
 connectToDB();