+++ src/models/address.js
@@ -10,7 +10,8 @@ const addressSchema = new Schema({
   street: { type: String, required: true },
   number: { type: Number, required: true },
   complement: { type: String },
-  postalCode: { type: String, required: true }
+  postalCode: { type: String, required: true },
+  userId: { type: Schema.Types.ObjectId, ref: 'users', required: true },
 });
 
 const Address = mongoose.model('addresses', addressSchema);